Fixed build failure when using Eclipse to build the Quaere module.

+++ Quaere/src/test/java/org/quaere/DSL_ElementOperatorsTest.java
@@ -1,11 +1,9 @@
 package org.quaere;
 
-import static org.quaere.DSL.*;
-
-import org.junit.Test;
 import junit.framework.Assert;
+import org.junit.Test;
+import static org.quaere.DSL.first;
 
-import java.util.List;
 import java.util.Arrays;
 
 public class DSL_ElementOperatorsTest {
@@ -17,11 +15,13 @@ public class DSL_ElementOperatorsTest {
     @Test
     public void canGetFirstFromIterable() {
         Iterable<Integer> numbers = Arrays.asList(5, 4, 1, 3, 9, 8, 6, 7, 2, 0);
-        Assert.assertEquals(5, (int) first(numbers));
+        // Note: Eclipse fails to build this test if the result from "first" is not casted to first's R.
+        Assert.assertEquals((Integer) 5, (Integer) first(numbers));
     }
     @Test
     public void canGetFirstFromQueryable() {
         Queryable<Integer> numbers = new QueryableIterable<Integer>(Arrays.asList(5, 4, 1, 3, 9, 8, 6, 7, 2, 0));
-        Assert.assertEquals(5, (int) first(numbers));
+        // Note: Eclipse fails to build this test if the result from "first" is not casted to first's R.
+        Assert.assertEquals((Integer) 5, (Integer) first(numbers));
     }
 }
